create shooter auto command
